+++ torch_geometric_signed_directed/nn/signed/__init__.py
@@ -1,1 +1,2 @@
-from .SSSNET import SIMPA, SSSNET
+from .SSSNET_node_clustering import SSSNET_node_clustering
+from .SIMPA import SIMPA
